requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
peewee==3.16.3
Pillow==10.0.1
pygame==2.5.2
